+++ Source/SelfService/Web/documentation/accessContainerRegistry.tsx
@@ -1,14 +1,23 @@
 // Copyright (c) Dolittle. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
-import React from 'react';
+import React, { useEffect, useState } from 'react';
 import ReactMarkdown from 'react-markdown';
 import gfm from 'remark-gfm';
 import { Info } from '../stores/documentationInfo';
 
+import { getContainerRegistry } from '../api/cicd';
+
+export type DockerCredentials = {
+    repoUrl: string
+};
+
+
+
 type Vars = {
     acrId: string
     subscriptionId: string
+    dockerCredentials: DockerCredentials
 };
 
 function template(vars: Vars): string {
@@ -29,6 +38,24 @@ az acr login -n ${vars.acrId} --subscription ${vars.subscriptionId}
 ~~~sh
 az acr repository list --name ${vars.acrId} -otable
 ~~~
+
+
+# Push images
+~~~sh
+# pull down your exmaple of choice
+# golang hello world
+git clone git@github.com:dolittle-entropy/go-hello-world
+
+# build go-hello-world docker image
+docker build -t go-hello-world .
+
+# tag the image with the url to container registry
+docker tag go-hello-world:latest ${vars.dockerCredentials.repoUrl}/go-hello-world:latest
+
+# push the image to container registry
+docker push ${vars.dockerCredentials.repoUrl}/go-hello-world:latest
+~~~
+
 `;
     return markdown.trim();
 };
@@ -41,9 +68,34 @@ export const Doc: React.FunctionComponent<Props> = (props) => {
     const _props = props!;
     const info = _props.info;
 
+    const [loaded, setLoaded] = useState(false);
+    const [containerRegistry, setContainerRegistry] = useState({});
+
+    useEffect(() => {
+        Promise.all([
+            getContainerRegistry(info.applicationId),
+        ]).then(values => {
+            setContainerRegistry(values[0]);
+            setLoaded(true);
+        });
+
+    }, []);
+
+    if (!loaded) {
+        return null;
+    }
+
+
+    const auths = JSON.parse(atob(containerRegistry['.dockerconfigjson'])).auths;
+    // Not great, but only one key for now
+    const repoUrl = Object.keys(auths)[0];
+    const credentials = auths[repoUrl] as DockerCredentials;
+    credentials.repoUrl = repoUrl;
+
     const vars = {
         acrId: info.containerRegistryName,
         subscriptionId: info.subscriptionId,
+        dockerCredentials: credentials
     } as Vars;
 
     const data = template(vars);